requirements.txt
httpx
requests
python-dotenv
anthropic
colorama
schedule 
requests